worker/requirements.txt
boto3==1.34.100
botocore==1.34.100
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
ffmpeg-python==0.2.0
mediapipe==0.10.11
numpy==1.26.4
opencv-python-headless==4.9.0.80
trimesh==4.3.2
pyopengl==3.1.7
pyrender==0.1.45
Pillow==10.3.0
scipy==1.13.0
scikit-learn==1.4.2
